archive: another fix for archives within archives...
Each archive will temporarily be extracted to it's archive member name,
and then moved to sub-archives/ (with a custom name based on the number
of sub-archives so far). When extracted, its contents will be placed
under a directory named after the archive member name.

As an example, if archive.cbz as the following contents:

sub.cbr
dir/sub.cbr

With both sub.cbr containing:

1.png
2.png

Then the temporary directory tree will look like this:

├── dir
│   └── sub.cbr
│       ├── 1.png
│       └── 2.png
├── sub-archives
│   ├── 0001.cbr
│   └── 0002.cbr
└── sub.cbr
    ├── 1.png
    └── 2.png

git-svn-id: svn://svn.code.sf.net/p/mcomix/code@1083 ef8490bc-f317-4725-81ed-e95a9aca1bcf