main: rework handling of "toggle" widgets
The "toggles" are the widgets that can be hidden/shown by the user:
menubar, toolbar, thumbnail bar, status bar, and scrolling bars.

Cleanup, factorize and simplify the code, as well as fixing the following bug:

When opening an empty archive, the OSD warning message can be scrolled
horizontally: the layout visible area size calculation was wrong; the
real visible state of each "toggle" widget must be used instead of the
corresponding preference state.

git-svn-id: svn://svn.code.sf.net/p/mcomix/code@1265 ef8490bc-f317-4725-81ed-e95a9aca1bcf