image_tools: fix load_pixbuf_size bug
Let GDK/PIL calculate the correct image dimensions after resizing,
otherwise the 2 compounded calculations can result in a rounding error,
making one of the dimension of the resulting pixbuf one pixel smaller
than what we asked for.

This in turn can result in some of the library covers being always
regenerated because the resulting thumbnails size does not match what
was asked for.

git-svn-id: svn://svn.code.sf.net/p/mcomix/code@1428 ef8490bc-f317-4725-81ed-e95a9aca1bcf